examples : add basic memory usage tracking to WhisperState.swift
This commit adds basic memory usage tracking to the `WhisperState.swift`
file in the Whisper SwiftUI demo.

The motivation for this to investigate a suspected memory leak reported
specifically when using whisper.cpp in Swift.

To run this code you need to have the XCFrameWork built and copy the
model and sample to use to the example:
```console
$ ./build-xcframework.sh
$ cp models/ggml-base-en.bin examples/whisper.swiftui/whisper.swiftui.demo/Resources/models
$ cp samples/jfk.wav examples/whisper.swiftui/whisper.swiftui.demo/Resources/samples
$ open examples/whisper.swiftui/whisper.swiftui.xcodeproj
```

The build in xcode and run the app and transcribe. In XCode the memory
can be inspected and is looks like memory is steadily increasing.