ci : update windows artifact uploads steps
This commit updates the windows job steps to use the built-in zip
feature of the actions/upload-artifact action. This simplifies the jobs
and also avoids the issue of zipping twice. This also modifies
windows-blas which was updated previously but to now make the zip achive
have an `Release` directory in it. Currently the files are all directly
in the archive which I think might be surprising to most developers.