mcp : add mcp client to example (wip)
I had misunderstood how the client server interaction works when using
the stdio transport protocol. As I understand itn now it is the client
process that spawns the server process and communicates with it via
pipes. The server process is not a long-running daemon, but rather a
temporary process that handles a single request and then exits.