vad : expose whisper_vad via C API
This commit exposes the whisper_vad function through the C API in
whisper.h.

The motivation for this is that currently the VAD functionality can be
used only by calling either whisper_full or whisper_full_parallel with
the `vad` parameter enabled. However, there might be use cases where
whisper_full_with_state is used directly in which case it is currently
not possible to use VAD.